wj_pytorch1.4_requirements.txt
albumentations==0.4.6
backcall==0.2.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2020.12.5
charset-normalizer==2.0.12
cleo==0.8.1
clikit==0.6.2
commonmark==0.9.1
crashtest==0.3.1
cycler==0.10.0
Cython @ file:///home/conda/feedstock_root/build_artifacts/cython_1602367281753/work
decorator==4.4.2
distlib==0.3.5
efficientnet-pytorch==0.7.1
filelock==3.4.1
flatbuffers==2.0
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.3
imgaug==0.4.0
importlib-resources==5.4.0
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jeepney==0.7.1
keyring==22.3.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.3.4
mkl-fft==1.2.1
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.4
munch==2.5.0
networkx==2.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
onnx==1.6.0
onnx-simplifier==0.3.3
onnxoptimizer==0.2.7
onnxruntime==1.10.0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
parso==0.7.1
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
platformdirs==2.4.0
poetry==1.1.14
poetry-core==1.0.8
pretrainedmodels==0.7.4
prompt-toolkit==3.0.29
protobuf==3.15.4
ptyprocess==0.7.0
Pygments==2.12.0
pylev==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==3.3.2
Shapely==1.8.1.post1
shellingham==1.4.0
six @ file:///home/linux1/recipes/ci/six_1610970791821/work
svgwrite==1.4.1
thop==0.0.31.post2005241907
tifffile==2020.9.3
timm==0.4.12
tomlkit==0.11.1
torch==1.4.0
torchvision @ file:///data2/wangj/software/torchvision-0.5.0%2Bcu100-cp36-cp36m-linux_x86_64.whl
tqdm==4.58.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.9
virtualenv==20.15.1
wcwidth==0.2.5
